finished migration to profiles. some clean-up

+++ src/main/java/com/m91snik/business/service/PaymentService.java
@@ -6,5 +6,5 @@ package com.m91snik.business.service;
 
 public interface PaymentService {
 
-    void doCredit(int amount);
+    void doCredit(long amount);
 }
